STYLE: Avoid old, C-style cast
Avoiding old-style casts on advice of:
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es48-avoid-casts
https://isocpp.github.io/CppCoreGuidelines/CppCoreGuidelines#es49-if-you-must-use-a-cast-use-a-named-cast
https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines-pro-type-cstyle-cast.html

+++ Examples/Numerics/FourierDescriptors1.cxx
@@ -162,8 +162,8 @@ main(int argc, char * argv[])
   //  Software Guide : EndLatex
 
   // Software Guide : BeginCodeSnippet
-  const auto powerOfTwo = (unsigned int)std::ceil(
-    std::log((double)(numberOfPoints)) / std::log((double)(2.0)));
+  const auto powerOfTwo = static_cast<unsigned int>(
+    std::ceil(std::log(static_cast<double>(numberOfPoints)) / std::log(2.0)));
 
   const unsigned int spectrumSize = 1 << powerOfTwo;
