+++ Modules/Numerics/Optimizers/src/itkOnePlusOneEvolutionaryOptimizer.cxx
@@ -27,7 +27,7 @@ OnePlusOneEvolutionaryOptimizer::OnePlusOneEvolutionaryOptimizer()
   m_MetricWorstPossibleValue = 0;
 
   m_Maximize = false;
-  m_Epsilon = (double)1.5e-4;
+  m_Epsilon = 1.5e-4;
   m_RandomGenerator = nullptr;
 
   m_Initialized = false;
